requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.6
aiosignal==1.3.2
attrs==25.3.0
cachetools==6.0.0
d20==1.1.2
discord.py==2.5.2
dnspython==2.7.0
frozenlist==1.6.0
idna==3.10
lark-parser==0.9.0
multidict==6.4.4
propcache==0.3.1
pymongo==4.13.0
python-dotenv==1.1.0
yarl==1.20.0
